Safari-609.1.11.1/LayoutTests/css3/scroll-snap/scroll-snap-inherit-expected.txt

Tests 'inherit' on scroll snap properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS noInherit.style['scroll-snap-type'] is ""
PASS window.getComputedStyle(noInherit).getPropertyValue('scroll-snap-type') is "both mandatory"
PASS noInherit.style['scroll-snap-align'] is ""
PASS window.getComputedStyle(noInherit).getPropertyValue('scroll-snap-align') is "center center"
PASS noInherit.style['scroll-padding'] is ""
PASS noInherit.style['scroll-padding-left'] is ""
PASS noInherit.style['scroll-padding-right'] is ""
PASS noInherit.style['scroll-padding-top'] is ""
PASS noInherit.style['scroll-padding-bottom'] is ""
PASS window.getComputedStyle(noInherit).getPropertyValue('scroll-padding') is "20px"
PASS noInherit.style['scroll-snap-margin'] is ""
PASS noInherit.style['scroll-snap-margin-left'] is ""
PASS noInherit.style['scroll-snap-margin-right'] is ""
PASS noInherit.style['scroll-snap-margin-top'] is ""
PASS noInherit.style['scroll-snap-margin-bottom'] is ""
PASS window.getComputedStyle(noInherit).getPropertyValue('scroll-snap-margin') is "0px 1px"

PASS inheritType.style['scroll-snap-type'] is "inherit"
PASS window.getComputedStyle(inheritType).getPropertyValue('scroll-snap-type') is "x proximity"
PASS inheritType.style['scroll-snap-align'] is ""
PASS window.getComputedStyle(inheritType).getPropertyValue('scroll-snap-align') is "center center"
PASS inheritType.style['scroll-padding'] is ""
PASS inheritType.style['scroll-padding-left'] is ""
PASS inheritType.style['scroll-padding-right'] is ""
PASS inheritType.style['scroll-padding-top'] is ""
PASS inheritType.style['scroll-padding-bottom'] is ""
PASS window.getComputedStyle(inheritType).getPropertyValue('scroll-padding') is "20px"
PASS inheritType.style['scroll-snap-margin'] is ""
PASS inheritType.style['scroll-snap-margin-left'] is ""
PASS inheritType.style['scroll-snap-margin-right'] is ""
PASS inheritType.style['scroll-snap-margin-top'] is ""
PASS inheritType.style['scroll-snap-margin-bottom'] is ""
PASS window.getComputedStyle(inheritType).getPropertyValue('scroll-snap-margin') is "0px 1px"

PASS inheritAlign.style['scroll-snap-type'] is ""
PASS window.getComputedStyle(inheritAlign).getPropertyValue('scroll-snap-type') is "both mandatory"
PASS inheritAlign.style['scroll-snap-align'] is "inherit"
PASS window.getComputedStyle(inheritAlign).getPropertyValue('scroll-snap-align') is "start end"
PASS inheritAlign.style['scroll-padding'] is ""
PASS inheritAlign.style['scroll-padding-left'] is ""
PASS inheritAlign.style['scroll-padding-right'] is ""
PASS inheritAlign.style['scroll-padding-top'] is ""
PASS inheritAlign.style['scroll-padding-bottom'] is ""
PASS window.getComputedStyle(inheritAlign).getPropertyValue('scroll-padding') is "20px"
PASS inheritAlign.style['scroll-snap-margin'] is ""
PASS inheritAlign.style['scroll-snap-margin-left'] is ""
PASS inheritAlign.style['scroll-snap-margin-right'] is ""
PASS inheritAlign.style['scroll-snap-margin-top'] is ""
PASS inheritAlign.style['scroll-snap-margin-bottom'] is ""
PASS window.getComputedStyle(inheritAlign).getPropertyValue('scroll-snap-margin') is "0px 1px"

PASS inheritPadding.style['scroll-snap-type'] is ""
PASS window.getComputedStyle(inheritPadding).getPropertyValue('scroll-snap-type') is "both mandatory"
PASS inheritPadding.style['scroll-snap-align'] is ""
PASS window.getComputedStyle(inheritPadding).getPropertyValue('scroll-snap-align') is "center center"
PASS inheritPadding.style['scroll-padding'] is "inherit"
PASS inheritPadding.style['scroll-padding-left'] is "inherit"
PASS inheritPadding.style['scroll-padding-right'] is "inherit"
PASS inheritPadding.style['scroll-padding-top'] is "inherit"
PASS inheritPadding.style['scroll-padding-bottom'] is "inherit"
PASS window.getComputedStyle(inheritPadding).getPropertyValue('scroll-padding') is "10%"
PASS inheritPadding.style['scroll-snap-margin'] is ""
PASS inheritPadding.style['scroll-snap-margin-left'] is ""
PASS inheritPadding.style['scroll-snap-margin-right'] is ""
PASS inheritPadding.style['scroll-snap-margin-top'] is ""
PASS inheritPadding.style['scroll-snap-margin-bottom'] is ""
PASS window.getComputedStyle(inheritPadding).getPropertyValue('scroll-snap-margin') is "0px 1px"

PASS inheritMargin.style['scroll-snap-type'] is ""
PASS window.getComputedStyle(inheritMargin).getPropertyValue('scroll-snap-type') is "both mandatory"
PASS inheritMargin.style['scroll-snap-align'] is ""
PASS window.getComputedStyle(inheritMargin).getPropertyValue('scroll-snap-align') is "center center"
PASS inheritMargin.style['scroll-padding'] is ""
PASS inheritMargin.style['scroll-padding-left'] is ""
PASS inheritMargin.style['scroll-padding-right'] is ""
PASS inheritMargin.style['scroll-padding-top'] is ""
PASS inheritMargin.style['scroll-padding-bottom'] is ""
PASS window.getComputedStyle(inheritMargin).getPropertyValue('scroll-padding') is "20px"
PASS inheritMargin.style['scroll-snap-margin'] is "inherit"
PASS inheritMargin.style['scroll-snap-margin-left'] is "inherit"
PASS inheritMargin.style['scroll-snap-margin-right'] is "inherit"
PASS inheritMargin.style['scroll-snap-margin-top'] is "inherit"
PASS inheritMargin.style['scroll-snap-margin-bottom'] is "inherit"
PASS window.getComputedStyle(inheritMargin).getPropertyValue('scroll-snap-margin') is "5px 5px 10px 10px"

PASS inheritAll.style['scroll-snap-type'] is "inherit"
PASS window.getComputedStyle(inheritAll).getPropertyValue('scroll-snap-type') is "x proximity"
PASS inheritAll.style['scroll-snap-align'] is "inherit"
PASS window.getComputedStyle(inheritAll).getPropertyValue('scroll-snap-align') is "start end"
PASS inheritAll.style['scroll-padding'] is "inherit"
PASS inheritAll.style['scroll-padding-left'] is "inherit"
PASS inheritAll.style['scroll-padding-right'] is "inherit"
PASS inheritAll.style['scroll-padding-top'] is "inherit"
PASS inheritAll.style['scroll-padding-bottom'] is "inherit"
PASS window.getComputedStyle(inheritAll).getPropertyValue('scroll-padding') is "10%"
PASS inheritAll.style['scroll-snap-margin'] is "inherit"
PASS inheritAll.style['scroll-snap-margin-left'] is "inherit"
PASS inheritAll.style['scroll-snap-margin-right'] is "inherit"
PASS inheritAll.style['scroll-snap-margin-top'] is "inherit"
PASS inheritAll.style['scroll-snap-margin-bottom'] is "inherit"
PASS window.getComputedStyle(inheritAll).getPropertyValue('scroll-snap-margin') is "5px 5px 10px 10px"

PASS inheritNone.style['scroll-snap-type'] is ""
PASS window.getComputedStyle(inheritNone).getPropertyValue('scroll-snap-type') is "none"
PASS inheritNone.style['scroll-snap-align'] is ""
PASS window.getComputedStyle(inheritNone).getPropertyValue('scroll-snap-align') is "none none"
PASS inheritNone.style['scroll-padding'] is ""
PASS inheritNone.style['scroll-padding-left'] is ""
PASS inheritNone.style['scroll-padding-right'] is ""
PASS inheritNone.style['scroll-padding-top'] is ""
PASS inheritNone.style['scroll-padding-bottom'] is ""
PASS window.getComputedStyle(inheritNone).getPropertyValue('scroll-padding') is "0px"
PASS inheritNone.style['scroll-snap-margin'] is ""
PASS inheritNone.style['scroll-snap-margin-left'] is ""
PASS inheritNone.style['scroll-snap-margin-right'] is ""
PASS inheritNone.style['scroll-snap-margin-top'] is ""
PASS inheritNone.style['scroll-snap-margin-bottom'] is ""
PASS window.getComputedStyle(inheritNone).getPropertyValue('scroll-snap-margin') is "0px"
PASS successfullyParsed is true

TEST COMPLETE
